requirements.txt
# Core Framework
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langgraph>=0.0.30

# LLM Providers
openai>=1.3.0

# Data Manipulation
pandas>=2.0.0
numpy>=1.24.0
ydata-profiling>=4.0.0  # For the ingestion agent's profile

# Visualization & Output
matplotlib>=3.7.0
seaborn>=0.12.0
tabulate>=0.9.0

# CLI Framework
click>=8.1.0

# Environment Management
python-dotenv>=1.0.0

# Optional: For more advanced data cleaning
scikit-learn>=1.2.0